+++ src/components/mui/formik-inputs/company-input-mui.js
@@ -16,25 +16,27 @@ import PropTypes from "prop-types";
 import TextField from "@mui/material/TextField";
 import Autocomplete from "@mui/material/Autocomplete";
 import CircularProgress from "@mui/material/CircularProgress";
-import { querySponsorships } from "../../actions/sponsorship-actions";
-import { DEBOUNCE_WAIT_250 } from "../../utils/constants";
+import ExpandMoreIcon from "@mui/icons-material/ExpandMore";
+import { useField } from "formik";
+import { queryCompanies } from "../../../actions/company-actions";
+import { DEBOUNCE_WAIT_250 } from "../../../utils/constants";
 
-const SponsorshipTypeInputMUI = ({
+const CompanyInputMUI = ({
   id,
   name,
-  formik,
   placeholder,
   plainValue,
   isMulti = false,
   ...rest
 }) => {
+  const [field, meta, helpers] = useField(name);
   const [options, setOptions] = useState([]);
   const [open, setOpen] = useState(false);
   const [inputValue, setInputValue] = useState("");
   const [loading, setLoading] = useState(false);
 
-  const value = formik?.values[name];
-  const error = formik?.touched[name] && formik?.errors[name];
+  const { value } = field;
+  const error = meta.touched && meta.error;
 
   const fetchOptions = async (input) => {
     if (!input) {
@@ -50,7 +52,7 @@ const SponsorshipTypeInputMUI = ({
         label: r.name
       }));
 
-    await querySponsorships(input, (results) => {
+    await queryCompanies(input, (results) => {
       setOptions(normalize(results));
       setLoading(false);
     });
@@ -98,7 +100,7 @@ const SponsorshipTypeInputMUI = ({
         : { id: parseInt(newValue.value), name: newValue.label };
     }
 
-    formik.setFieldValue(name, theValue);
+    helpers.setValue(theValue);
   };
 
   return (
@@ -117,9 +119,11 @@ const SponsorshipTypeInputMUI = ({
       onChange={handleChange}
       loading={loading}
       fullWidth
+      popupIcon={<ExpandMoreIcon />}
       renderInput={(params) => (
         <TextField
           {...params}
+          name={field.name}
           placeholder={placeholder}
           variant="outlined"
           error={Boolean(error)}
@@ -147,13 +151,12 @@ const SponsorshipTypeInputMUI = ({
   );
 };
 
-SponsorshipTypeInputMUI.propTypes = {
+CompanyInputMUI.propTypes = {
   id: PropTypes.string,
   name: PropTypes.string.isRequired,
-  formik: PropTypes.object.isRequired,
   placeholder: PropTypes.string,
   plainValue: PropTypes.bool,
   isMulti: PropTypes.bool
 };
 
-export default SponsorshipTypeInputMUI;
+export default CompanyInputMUI;
