+++ src/components/mui/table/mui-table.js
@@ -22,7 +22,7 @@ import {
   FIFTY_PER_PAGE,
   TWENTY_PER_PAGE
 } from "../../../utils/constants";
-import showConfirmDialog from "../components/showConfirmDialog";
+import showConfirmDialog from "../showConfirmDialog";
 
 const MuiTable = ({
   columns = [],
