+++ src/pages/sponsors/sponsor-form-item-list-page/components/item-form.js
@@ -14,11 +14,11 @@ import T from "i18n-react";
 import * as yup from "yup";
 import { FormikProvider, useFormik } from "formik";
 import { addIssAfterDateFieldValidator } from "../../../../utils/yup";
-import MuiFormikTextField from "../../../../components/inputs/mui-formik-textfield";
+import MuiFormikTextField from "../../../../components/mui/formik-inputs/mui-formik-textfield";
 import AdditionalInputList from "../../components/additional-input-list";
 import useScrollToError from "../../../../hooks/useScrollToError";
-import MuiFormikUpload from "../../../../components/inputs/mui-formik-upload";
-import MuiFormikPriceField from "../../../../components/inputs/mui-formik-pricefield";
+import MuiFormikUpload from "../../../../components/mui/formik-inputs/mui-formik-upload";
+import MuiFormikPriceField from "../../../../components/mui/formik-inputs/mui-formik-pricefield";
 import FormikTextEditor from "../../../../components/inputs/formik-text-editor";
 
 const buildInitialValues = (data) => {
