+++ src/pages/sponsors/sponsor-forms-list-page/components/form-template/form-template-form.js
@@ -14,9 +14,9 @@ import T from "i18n-react";
 import * as yup from "yup";
 import { FormikProvider, useFormik } from "formik";
 import { addIssAfterDateFieldValidator } from "../../../../../utils/yup";
-import DropdownCheckbox from "../../../../../components/mui/components/dropdown-checkbox";
-import MuiFormikTextField from "../../../../../components/inputs/mui-formik-textfield";
-import MuiFormikDatepicker from "../../../../../components/inputs/mui-formik-datepicker";
+import DropdownCheckbox from "../../../../../components/mui/dropdown-checkbox";
+import MuiFormikTextField from "../../../../../components/mui/formik-inputs/mui-formik-textfield";
+import MuiFormikDatepicker from "../../../../../components/mui/formik-inputs/mui-formik-datepicker";
 import AdditionalInputList from "../../../components/additional-input-list";
 import useScrollToError from "../../../../../hooks/useScrollToError";
 import FormikTextEditor from "../../../../../components/inputs/formik-text-editor";
