+++ src/pages/sponsors/sponsor-forms-list-page/components/global-template/select-templates-dialog.js
@@ -16,7 +16,7 @@ import {
   Typography
 } from "@mui/material";
 import CloseIcon from "@mui/icons-material/Close";
-import SearchInput from "../../../../../components/mui/components/search-input";
+import SearchInput from "../../../../../components/mui/search-input";
 import { getGlobalTemplates } from "../../../../../actions/sponsor-forms-actions";
 import { DEFAULT_PER_PAGE } from "../../../../../utils/constants";
 import MuiInfiniteTable from "../../../../../components/mui/infinite-table";
@@ -70,7 +70,7 @@ const SelectTemplatesDialog = ({
 
   const handleOnSave = () => {
     onSave(selectedRows);
-  }
+  };
 
   const columns = [
     {
