+++ src/pages/sponsors_inventory/popup/add-form-template-item-popup.js
@@ -22,7 +22,7 @@ import SearchIcon from "@mui/icons-material/Search";
 import CloseIcon from "@mui/icons-material/Close";
 import ImageIcon from "@mui/icons-material/Image";
 import MuiTable from "../../../components/mui/table/mui-table";
-import MenuButton from "../../../components/mui/components/menu-button";
+import MenuButton from "../../../components/mui/menu-button";
 import {
   clearAllSelectedInventoryItems,
   getInventoryItems,
