+++ src/pages/sponsors_inventory/popup/form-template-from-duplicate-popup.js
@@ -20,7 +20,7 @@ import SwapVertIcon from "@mui/icons-material/SwapVert";
 import SearchIcon from "@mui/icons-material/Search";
 import CloseIcon from "@mui/icons-material/Close";
 import MuiTable from "../../../components/mui/table/mui-table";
-import MenuButton from "../../../components/mui/components/menu-button";
+import MenuButton from "../../../components/mui/menu-button";
 
 const FormTemplateFromDuplicateDialog = ({
   open,
